Expose getFilenameAndExtension()
