+++ util/index.js
@@ -25,6 +25,7 @@ export { default as flatten } from './flatten'
 export { default as flattenOften } from './flattenOften'
 export { default as forEach } from './forEach'
 export { default as getKeyForPath } from './getKeyForPath'
+export { default as getFilenameAndExtension } from './getFilenameAndExtension'
 export { default as getRangeFromMatrix } from './getRangeFromMatrix'
 export { default as getRelativeBoundingRect } from './getRelativeBoundingRect'
 export { default as getRelativeMouseBounds } from './getRelativeMouseBounds'
